Refactor for structured streaming
