+++ sparkbench/pom.xml
@@ -99,6 +99,13 @@
     </profile>
 
     <profile>
+      <id>structuredStreaming</id>
+      <modules>
+        <module>structuredStreaming</module>
+      </modules>
+    </profile>
+
+    <profile>
       <id>allModules</id>
       <modules>
         <module>micro</module>
@@ -121,6 +128,15 @@
         <spark.version>2.0.0</spark.version>
         <spark.bin.version>2.0</spark.bin.version>
       </properties>
+      <modules>
+        <module>micro</module>
+        <module>ml</module>
+        <module>websearch</module>
+        <module>graph</module>
+        <module>sql</module>
+        <module>streaming</module>
+        <module>structuredStreaming</module>
+      </modules>
       <activation>
         <property>
           <name>!spark</name>
@@ -148,6 +164,15 @@
         <spark.version>2.0.0</spark.version>
         <spark.bin.version>2.0</spark.bin.version>
       </properties>
+      <modules>
+        <module>micro</module>
+        <module>ml</module>
+        <module>websearch</module>
+        <module>graph</module>
+        <module>sql</module>
+        <module>streaming</module>
+        <module>structuredStreaming</module>
+      </modules>
       <activation>
         <property>
           <name>spark</name>
