+++ sparkbench/structuredStreaming/pom.xml
@@ -10,9 +10,9 @@
   </parent>
 
   <groupId>com.intel.hibench.sparkbench</groupId>
-  <artifactId>sparkbench-streaming</artifactId>
+  <artifactId>sparkbench-structuredStreaming</artifactId>
   <packaging>jar</packaging>
-  <name>sparkbench-streaming</name>
+  <name>sparkbench-structuredStreaming</name>
 
   <dependencies>
     <dependency>
@@ -32,6 +32,12 @@
       <version>${spark.version}</version>
       <scope>provided</scope>
     </dependency>
+    <dependency>
+      <groupId>org.apache.spark</groupId>
+      <artifactId>spark-sql_${scala.binary.version}</artifactId>
+      <version>${spark.version}</version>
+      <scope>provided</scope>
+    </dependency>
   </dependencies>
 
   <profiles>
@@ -40,30 +46,19 @@
       <dependencies>
         <dependency>
           <groupId>org.apache.spark</groupId>
-          <artifactId>spark-streaming-kafka_${scala.binary.version}</artifactId>
-          <version>1.6.0</version>
+          <artifactId>spark-streaming-kafka-0-8_2.11</artifactId>
+          <version>2.0.0</version>
         </dependency>
-      </dependencies>
-      <activation>
-        <property>
-          <name>!spark</name>
-        </property>
-      </activation>
-    </profile>
-
-    <profile>
-      <id>spark1.6</id>
-      <dependencies>
         <dependency>
           <groupId>org.apache.spark</groupId>
-          <artifactId>spark-streaming-kafka_${scala.binary.version}</artifactId>
-          <version>1.6.0</version>
+          <artifactId>spark-sql-kafka-0-10_2.11</artifactId>
+          <version>2.0.2</version>
         </dependency>
       </dependencies>
       <activation>
         <property>
           <name>spark</name>
-          <value>1.6</value>
+          <value>2.0</value>
         </property>
       </activation>
     </profile>
@@ -89,23 +84,5 @@
         </property>
       </activation>
     </profile>
-
-    <profile>
-      <id>spark2.1</id>
-      <dependencies>
-        <dependency>
-          <groupId>org.apache.spark</groupId>
-          <artifactId>spark-streaming-kafka-0-8_2.11</artifactId>
-          <version>2.1.0</version>
-        </dependency>
-      </dependencies>
-      <activation>
-        <property>
-          <name>spark</name>
-          <value>2.1</value>
-        </property>
-      </activation>
-    </profile>
-
   </profiles>
 </project>
