+++ sparkbench/structuredStreaming/src/main/resource/config.properties
@@ -1,0 +1,1 @@
+tupleCount=1000000
