+++ sparkbench/structuredStreaming/src/main/scala/com/intel/hibench/sparkbench/structuredstreaming/application/StructuredWordCount.scala
@@ -15,18 +15,16 @@
  * limitations under the License.
  */
 
-package com.intel.hibench.sparkbench.streaming.application
+package com.intel.hibench.sparkbench.structuredstreaming.application
 
-import com.intel.hibench.common.streaming.metrics.KafkaReporter
-import com.intel.hibench.sparkbench.streaming.util.SparkBenchConfig
+import com.intel.hibench.common.streaming.UserVisitParser
+import com.intel.hibench.sparkbench.structuredstreaming.util.SparkBenchConfig
 import org.apache.spark.sql.Column
 import org.apache.spark.sql.DataFrame
-import org.apache.spark.sql.ForeachWriter
-import org.apache.spark.sql.Row
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.functions._
 
-class StructuredRepartition() extends StructuredBenchBase {
+class StructuredWordCount() extends StructuredBenchBase {
 
   override def process(ds: DataFrame, config: SparkBenchConfig) = {
 
@@ -34,28 +32,17 @@ class StructuredRepartition() extends StructuredBenchBase {
     val spark = SparkSession.builder.appName("structured " + config.benchName).getOrCreate()
     import spark.implicits._
 
-    val results = ds.repartition(config.coreNumber)
+    // Project Line to UserVisit
+    val words = ds.map(row => {
+      val userVisit = UserVisitParser.parse(row.getAs[String]("value"))
+      userVisit.getIp
+    })
+
+    val wordCounts = words.groupBy("value").count()
     
-    val query = results.writeStream
+    val query = wordCounts.writeStream
+      .outputMode("complete")
       .format("console")
-      .foreach(new ForeachWriter[Row] {
-        var reporter: KafkaReporter = _
-
-        def open(partitionId: Long, version: Long): Boolean = {
-          val reportTopic = config.reporterTopic
-          val brokerList = config.brokerList
-          reporter = new KafkaReporter(reportTopic, brokerList)
-          true
-        }
-
-        def close(errorOrNull: Throwable): Unit = {}
-
-        def process(record: Row): Unit = {
-          val inTime = record(0).asInstanceOf[String].toLong
-          val outTime = System.currentTimeMillis()
-          reporter.report(inTime, outTime)
-        }
-      })
       .start()
 
     query.awaitTermination()
